+++ src/test/scala/saferis/tests/DataManipulationLayerSpecs.scala
@@ -4,6 +4,7 @@ import saferis.*
 import saferis.dml.*
 import zio.*
 import zio.test.*
+import PostgresTestContainer.DataSourceProvider
 
 object DataManipulationLayerSpecs extends ZIOSpecDefault:
   val xaLayer = DataSourceProvider.default >>> Transactor.default
