+++ src/test/scala/saferis/tests/TableQuerySpecs.scala
@@ -3,6 +3,7 @@ import saferis.*
 import saferis.dml.*
 import zio.*
 import zio.test.*
+import PostgresTestContainer.DataSourceProvider
 
 object TableQuerySpecs extends ZIOSpecDefault:
   val xaLayer = DataSourceProvider.default >>> Transactor.default
